split test_query.cpp in half
